Move FaultyComponent into chiffre package
